added code to access review table in datastore

+++ src/main/java/com/google/codeu/data/AverageRating.java
@@ -39,7 +39,7 @@ public class AverageRating {
     return restaurant;
   }
   
-  public long getAvgRating() {
+  public double getAvgRating() {
     return avgRating;
   }
